system_tracking_orders/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.12
django-crontab==0.7.1
djangorestframework==3.16.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
